fix: remove ads-old TextInput
