+++ app/client/src/components/utils/ReduxFormTextField.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import type { WrappedFieldMetaProps, WrappedFieldInputProps } from "redux-form";
 import { Field } from "redux-form";
-import type { InputType } from "@appsmith/ads-old";
 import { Input, NumberInput } from "@appsmith/ads";
 
 import type { Intent } from "constants/DefaultTheme";
@@ -48,7 +47,7 @@ export interface FormTextFieldProps {
   name: string;
   placeholder: string;
   description?: string;
-  type?: InputType;
+  type?: "text" | "password" | "number" | "email" | "tel";
   label?: React.ReactNode;
   intent?: Intent;
   disabled?: boolean;
